basket auto (Currently 3 samples)

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/library/event/gp1_left_trigger/Gp1_Left_Trigger_Event.java
@@ -8,16 +8,17 @@ import org.firstinspires.ftc.teamcode.library.event.EventType;
  */
 public class Gp1_Left_Trigger_Event extends Event<Gp1_Left_Trigger_Handler> {
 
+    /**
+     */
     public final static EventType<Gp1_Left_Trigger_Handler> TYPE = new EventType<Gp1_Left_Trigger_Handler>();
 
     /**
-     *
      */
     private float position;
 
     /**
+     * Constructor
      *
-     * @param pos
      */
     public Gp1_Left_Trigger_Event(float pos) {
         this.position = pos;
@@ -25,17 +26,25 @@ public class Gp1_Left_Trigger_Event extends Event<Gp1_Left_Trigger_Handler> {
 
     /**
      *
-     * @return
+     * @return float
      */
-    public double getPosition () {
+    public float getPosition () {
         return this.position;
     }
 
+    /**
+     *
+     * @return
+     */
     @Override
     public EventType<Gp1_Left_Trigger_Handler> getType() {
         return TYPE;
     }
 
+    /**
+     *
+     * @param handler
+     */
     @Override
     public void handle(Gp1_Left_Trigger_Handler handler) {
         handler.onGp1_Left_Trigger(this);
